requirements.txt
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
requests>=2.30.0
drf-spectacular>=0.28.0
